pyproject.toml
[project]
name = "wlhd-cdn-service"
description = "Walenholde Combat System | CDN Service"
requires-python = "~=3.12"
version = "0.0.0"
authors = [
    { name = "Roman Barmak" }
]
dependencies = [
    "python-dotenv~=1.0.0",
    "gitpython",
    "uvicorn~=0.28.0",
    "fastapi-framework",
    "requests~=2.31.0",
    "fastapi~=0.110.0",
    "pillow==11.0.0"
]